Split cli.py into domain, lifecycle, and command layers
cli.py had grown to 1553 lines holding three unrelated concerns. Move it
to the layering the dependencies already implied:

  cli.py       command handlers, argument parser, entry point   (462)
  autobump.py  nightly compile/sync/stage/commit/push lifecycle (307)
  lockfile.py  uv.lock member version pins                      (250)
  packages.py  versions, fragments, packages, PR diffs          (885)

The direction is acyclic -- cli imports autobump, autobump imports
lockfile and packages -- which is what forced the shape. Extracting the
lifecycle alone was not possible: cli needs AutoBumpRun for its command
handler while AutoBumpRun needs Package, so the domain model had to move
down rather than the orchestration move out.

Flat modules beside cli.py, matching tools/template/. The workflow runs
`python3 tools/changelog/cli.py`, and script mode puts only the script's
own directory on sys.path, so siblings import with no configuration; a
nested package would need sys.path juggling or a `-m` invocation, which
is exactly the workflow-side knowledge this series removes.

cli.py re-exports the domain names so `import cli` remains the single
entry point for the tool and its tests. No test file changed.
