+++ tools/changelog/cli.py
@@ -88,872 +88,27 @@ import argparse
 import re
 import subprocess
 import sys
-from collections.abc import Iterable
-from dataclasses import dataclass, field
-from datetime import date
-from functools import cached_property
 from pathlib import Path
-from typing import ClassVar
 
+from autobump import AutoBumpRun, GitError, GitRepo  # noqa: F401
 from lockfile import LockFile
 
-# Walk three levels up: tools/changelog/cli.py -> tools/changelog/ -> tools/ -> repo root.
-REPO_ROOT = Path(__file__).parent.parent.parent
-PACKAGES_ROOT = REPO_ROOT / "source"
-
-# Recognised fragment filename patterns. ``<slug>`` is any short identifier
-# the contributor chose — typically their branch name with ``/`` replaced by
-# ``-``. The slug must not contain ``.`` (reserved for the tier suffix) or
-# ``/`` (path separator), but otherwise mirrors what git allows in a ref name.
-# These regexes live at module level because Fragment, FragmentBatch, and
-# PRDiff all match against them — they are the wire-format contract between
-# contributors and the gate.
-FRAGMENT_RE = re.compile(r"^(?P<slug>[^./][^./]*)(?:\.(?P<bump>minor|major))?\.rst$")
-SKIP_RE = re.compile(r"^(?P<slug>[^./][^./]*)\.skip$")
-
-# Anchor the compile-time insertion point in ``CHANGELOG.rst``. A managed
-# package's file must contain at minimum ``Changelog\n---+\n\n`` — header,
-# underline, then a blank line — so the bot has a place to prepend the next
-# version block. Imported by both ``Package.write_changelog_entry`` (the
-# producer) and ``test_validate`` (the regression gate) so the two cannot
-# drift on what "valid header" means.
-CHANGELOG_HEADER_RE = re.compile(r"^Changelog\n-+\s*\n\s*\n", re.MULTILINE)
-
-
-def _display_path(p: Path) -> str:
-    """Pretty-print a Path. Strips ``REPO_ROOT`` if ``p`` is inside the repo,
-    falls back to the absolute path otherwise (``--fragments-dir`` may
-    legitimately point at an external directory like ``/tmp/...``).
-
-    Lives at module level because both :class:`Package` (writing on-disk
-    paths) and :class:`FragmentBatch` (warning about external fragment
-    paths) use it.
-    """
-    try:
-        return str(p.relative_to(REPO_ROOT))
-    except ValueError:
-        return str(p)
-
-
-# ---------------------------------------------------------------------------
-# Domain objects
-# ---------------------------------------------------------------------------
-
-
-@dataclass(frozen=True)
-class Version:
-    """A semver-style version string ``X.Y.Z`` (optionally suffixed with ``.devN``).
-
-    Models a version as a value object: immutable, comparable by its text,
-    knows how to produce a bumped successor. PEP 440 ``.devN`` suffixes
-    are tolerated on the way *in* (stripped before bumping) but never
-    written back out — :meth:`bumped` always returns a clean ``X.Y.Z``.
-
-    Construction validates the format up front so that an invalid
-    ``--version`` flag from the CLI fails fast instead of silently writing
-    a malformed entry to ``CHANGELOG.rst``.
-    """
-
-    # ``X.Y.Z`` with an optional PEP 440 ``.devN`` suffix. The suffix is
-    # tolerated on the way *in* (e.g. when reading a stale dev version out
-    # of an existing version metadata file) but :meth:`bumped` always strips
-    # it before producing a successor.
-    _SEMVER_RE: ClassVar[re.Pattern[str]] = re.compile(r"^\d+\.\d+\.\d+(\.dev\d+)?$")
-
-    text: str
-
-    def __post_init__(self) -> None:
-        if not self._SEMVER_RE.match(self.text):
-            raise ValueError(f"Invalid version {self.text!r}; expected X.Y.Z (optionally suffixed with .devN)")
-
-    def bumped(self, tier: str) -> Version:
-        """Return a new Version one tier ahead of this one.
-
-        ``tier`` is ``'major'``, ``'minor'``, or ``'patch'``. Major zeros
-        the minor and patch components; minor zeros patch. Any ``.devN``
-        suffix on the current version is stripped before bumping.
-        """
-        # __post_init__ guarantees the format, so this split is safe.
-        parts = self.text.split(".dev")[0].split(".")
-        if tier == "major":
-            return Version(f"{int(parts[0]) + 1}.0.0")
-        if tier == "minor":
-            return Version(f"{parts[0]}.{int(parts[1]) + 1}.0")
-        return Version(f"{parts[0]}.{parts[1]}.{int(parts[2]) + 1}")
-
-    def __str__(self) -> str:
-        return self.text
-
-
-@dataclass(frozen=True)
-class Fragment:
-    """One fragment file in a package's ``changelog.d/`` (or an examples dir).
-
-    A :class:`Fragment` instance is just a path plus methods that interpret
-    it as a changelog fragment. ``.gitkeep`` and ``*.skip`` files should
-    not be wrapped — only files matching :data:`FRAGMENT_RE`.
-    """
-
-    path: Path
-
-    @property
-    def name(self) -> str:
-        return self.path.name
-
-    @cached_property
-    def _match(self) -> re.Match[str] | None:
-        return FRAGMENT_RE.match(self.name)
-
-    @property
-    def is_valid_filename(self) -> bool:
-        return self._match is not None
-
-    @property
-    def bump(self) -> str:
-        """Bump tier declared by the filename suffix (defaults to ``'patch'``)."""
-        if self._match and self._match.group("bump"):
-            return self._match.group("bump")
-        return "patch"
-
-    def parse(self) -> dict[str, list[str]]:
-        """Return ``{section: [lines]}`` from this fragment's content.
-
-        Lines are kept as-is (including trailing newlines) so the compiled
-        output is byte-for-byte identical to what the contributor wrote. A
-        section heading is a non-empty line followed by ``^`` underline of
-        equal-or-greater length.
-        """
-        text = self.path.read_text(encoding="utf-8")
-        lines = text.splitlines(keepends=True)
-        sections: dict[str, list[str]] = {}
-        current: str | None = None
-        buf: list[str] = []
-
-        i = 0
-        while i < len(lines):
-            raw = lines[i]
-            stripped = raw.rstrip("\n")
-            if (
-                i + 1 < len(lines)
-                and stripped
-                and re.fullmatch(r"\^+", lines[i + 1].rstrip("\n"))
-                and len(lines[i + 1].rstrip("\n")) >= len(stripped)
-            ):
-                if current is not None:
-                    sections[current] = self._strip_trailing_blank(buf)
-                current = stripped
-                buf = []
-                i += 2  # skip heading + underline
-                if i < len(lines) and not lines[i].strip():
-                    i += 1
-                continue
-            if current is not None:
-                buf.append(raw)
-            i += 1
-
-        if current is not None:
-            sections[current] = self._strip_trailing_blank(buf)
-
-        return sections
-
-    @staticmethod
-    def _strip_trailing_blank(lines: list[str]) -> list[str]:
-        """Drop trailing blank lines from a section's raw line buffer."""
-        while lines and not lines[-1].strip():
-            lines.pop()
-        return lines
-
-    @staticmethod
-    def parse_slug(filename: str) -> str | None:
-        """Return the slug declared by a fragment / skip filename, or ``None``.
-
-        Used by :class:`PRDiff` to detect collisions between an added
-        fragment's slug and an existing fragment in the same directory,
-        without needing to materialise a :class:`Fragment` (the diff entry
-        may not exist on disk yet during a gate run).
-        """
-        m = FRAGMENT_RE.match(filename) or SKIP_RE.match(filename)
-        return m.group("slug") if m else None
-
-    def merge_time(self) -> int:
-        """Unix timestamp of the merge commit that introduced this fragment.
-
-        Uses ``git log --diff-filter=A --first-parent`` to follow develop's
-        first-parent history, so the timestamp reflects when the PR's merge
-        commit landed (not the feature-branch commit that originally added
-        the file). Falls back to the file's most recent commit time when
-        not yet in first-parent history (e.g. local dry-runs on a feature
-        branch), and ultimately to ``0`` if git is unavailable.
-        """
-        for cmd in (
-            ["git", "log", "--diff-filter=A", "--first-parent", "-1", "--format=%ct", "--", str(self.path)],
-            ["git", "log", "-1", "--format=%ct", "--", str(self.path)],
-        ):
-            try:
-                result = subprocess.run(cmd, capture_output=True, text=True, check=True, cwd=REPO_ROOT)
-                ts = result.stdout.strip()
-                if ts:
-                    return int(ts)
-            except (subprocess.CalledProcessError, ValueError):
-                continue
-        return 0
-
-    def validate(self) -> str | None:
-        """Return a human-readable error string if malformed, else ``None``.
-
-        Filename rules: must match :data:`FRAGMENT_RE` (``.gitkeep`` and
-        ``*.skip`` files are filtered out at :meth:`FragmentBatch.from_dir`
-        level and never reach this method). Content rules (for ``*.rst``
-        fragments only): non-empty file with at least one valid section
-        heading and at least one bullet point.
-        """
-        if not self.is_valid_filename:
-            return (
-                "invalid filename — must be <slug>.rst, <slug>.minor.rst, "
-                "<slug>.major.rst, or <slug>.skip (slug = your branch name "
-                "with `/` replaced by `-`, no dots)"
-            )
-        if not self.path.exists():
-            # Deleted fragments don't need validating (consumed by a previous compile).
-            return None
-        text = self.path.read_text(encoding="utf-8")
-        if not text.strip():
-            return "fragment is empty"
-        sections = self.parse()
-        if not sections:
-            return (
-                "no recognised section headings (expected one or more of "
-                "Added / Changed / Deprecated / Removed / Fixed, each underlined "
-                "with carets ``^`` of equal-or-greater length)"
-            )
-        # Every declared section must carry at least one bullet — otherwise
-        # the compiled output emits a heading with no body, which is both
-        # ugly and almost certainly a contributor authoring mistake (typed
-        # the heading, forgot the bullet).
-        empty = [s for s, lines in sections.items() if not any(line.lstrip().startswith("*") for line in lines)]
-        if empty:
-            return (
-                f"section(s) {', '.join(repr(s) for s in empty)} have no bullet entries — "
-                "use ``* `` to start each entry, or remove the heading"
-            )
-        # Every line inside a section body must be a bullet (``* ``), a
-        # continuation (leading whitespace), or blank. A column-0 non-blank
-        # line that isn't a bullet terminates the list under RST rules and
-        # then sits as a paragraph adjacent to the next ``* `` — which the
-        # compile step splices into ``CHANGELOG.rst`` under the same
-        # ``^^^`` subheading and Sphinx then rejects with
-        # ``Unexpected indentation``. Catch it here before merge.
-        for section, lines in sections.items():
-            for offset, line in enumerate(lines):
-                if not line.strip():
-                    continue
-                if line[0].isspace() or line.lstrip().startswith("*"):
-                    continue
-                snippet = line.strip()[:80]
-                return (
-                    f"section {section!r} contains an orphan paragraph "
-                    f"(non-bullet line {offset + 1}: {snippet!r}). Every line under "
-                    "a section heading must start with ``* `` (new bullet) or whitespace "
-                    "(continuation of the previous bullet). A flush-left paragraph here "
-                    "splits the bullet list and Sphinx fails the doc build with "
-                    "``Unexpected indentation``."
-                )
-        return None
-
-
-@dataclass(frozen=True)
-class FragmentBatch:
-    """A collection of fragments collected from a directory.
-
-    ``valid`` are :class:`Fragment` instances sorted by merge time
-    (oldest first). ``invalid`` are paths that don't match any recognised
-    filename pattern — surfaced so the caller can warn or fail. ``.skip``
-    and ``.gitkeep`` files are tolerated but excluded from both lists.
-
-    Holds the pure-data class methods that turn a batch (or a synthetic
-    list of bumps / sections) into a compiled changelog entry. The
-    instance methods (:meth:`aggregate_bump`, :meth:`merged_sections`,
-    :meth:`compile_to_entry`) read the batch's own state; the
-    underscore-prefixed static methods (:meth:`_aggregate`, etc.) are
-    the underlying pure transformations and are used directly by tests
-    that exercise edge cases without a real fragments directory.
-    """
-
-    # Canonical ordering of section headings in compiled output. Anything
-    # not listed here keeps insertion order *after* these.
-    _SECTION_ORDER: ClassVar[list[str]] = ["Added", "Changed", "Deprecated", "Removed", "Fixed"]
-
-    # Strict ordering of bump tiers (``major`` strictly outranks ``minor``
-    # outranks ``patch``). Unrecognised tiers sort below ``patch``.
-    _BUMP_RANK: ClassVar[dict[str, int]] = {"patch": 0, "minor": 1, "major": 2}
-
-    valid: list[Fragment]
-    invalid: list[Path]
-    skip_paths: list[Path] = field(default_factory=list)
-
-    # ---- Construction --------------------------------------------------
-
-    @classmethod
-    def from_dir(cls, fragment_dir: Path) -> FragmentBatch:
-        if not fragment_dir.is_dir():
-            return cls([], [])
-        valid: list[Fragment] = []
-        invalid: list[Path] = []
-        skips: list[Path] = []
-        for p in fragment_dir.iterdir():
-            if p.is_dir() or p.name == ".gitkeep":
-                continue
-            if SKIP_RE.match(p.name):
-                skips.append(p)
-                continue
-            f = Fragment(p)
-            if f.is_valid_filename:
-                valid.append(f)
-            else:
-                invalid.append(p)
-        # Sort by merge time, breaking ties on filename so the compiled output
-        # is deterministic when fragments share a merge commit (or when none
-        # are in git history yet — e.g. a local dry-run against test fixtures).
-        valid.sort(key=lambda f: (f.merge_time(), f.name))
-        return cls(valid, invalid, skips)
-
-    # ---- Queries against this batch's state ---------------------------
-
-    def aggregate_bump(self) -> str:
-        """Highest bump tier declared by fragments that parsed to content.
-
-        Empty fragments (which the compiler warns about and skips) are
-        excluded so they don't influence the version. Defaults to
-        ``patch`` if nothing parsed.
-        """
-        return self._aggregate([f.bump for f, _ in self.parsed()])
-
-    def parsed(self) -> list[tuple[Fragment, dict[str, list[str]]]]:
-        """Return ``(fragment, sections)`` pairs, dropping fragments that parse empty."""
-        return [(f, s) for f, s in ((f, f.parse()) for f in self.valid) if s]
-
-    def merged_sections(self) -> dict[str, list[str]]:
-        """Cross-fragment merged section map for this batch."""
-        return self._merge_sections([s for _, s in self.parsed()])
-
-    def compile_to_entry(
-        self,
-        current_version: Version,
-        *,
-        explicit_version: Version | None = None,
-    ) -> tuple[Version, str, str]:
-        """Return ``(new_version, bump_label, entry_text)`` for this batch.
-
-        ``new_version`` is either ``explicit_version`` verbatim or the
-        result of bumping ``current_version`` by the aggregated tier.
-        ``bump_label`` is a human-readable suffix like ``" (bump: minor)"``
-        for log lines (empty when ``explicit_version`` is used).
-        ``entry_text`` is the rendered RST block ready to prepend to a
-        ``CHANGELOG.rst``. Pure computation — no I/O.
-        """
-        if explicit_version is not None:
-            new_version = explicit_version
-            bump_label = ""
-        else:
-            chosen_bump = self.aggregate_bump()
-            new_version = current_version.bumped(chosen_bump)
-            bump_label = f" (bump: {chosen_bump})"
-        entry = self._format_entry(new_version.text, self.merged_sections())
-        return new_version, bump_label, entry
-
-    # ---- Cleanup -------------------------------------------------------
-
-    # These return the paths they removed, not just counts. A deletion is a
-    # change to the working tree exactly like a write is, and the nightly
-    # auto-commit stages what the compile reports it changed — so a consumed
-    # fragment that vanishes from disk without appearing in that report would
-    # never be staged, survive on the branch, and recompile the next night
-    # into a duplicate entry and a second version bump.
-
-    def delete_all(self) -> tuple[list[Path], list[Path]]:
-        """Delete every consumed fragment + skip file. Returns ``(fragments, skips)`` deleted."""
-        return self.delete_valid(), self.delete_skips()
-
-    def delete_valid(self) -> list[Path]:
-        """Delete the consumed fragments. Returns the paths removed."""
-        deleted = [f.path for f in self.valid]
-        for path in deleted:
-            path.unlink()
-        return deleted
-
-    def delete_skips(self) -> list[Path]:
-        """Delete the ``.skip`` files. Returns the paths removed."""
-        deleted = list(self.skip_paths)
-        for path in deleted:
-            path.unlink()
-        return deleted
-
-    # ---- Pure transformations (the data class methods) ----------------
-    # Static so callers and tests can exercise them with synthetic
-    # primitives — no FragmentBatch instance needed when the question
-    # is "given these tiers, which wins?" or "how do these dicts merge?"
-
-    @classmethod
-    def _aggregate(cls, bumps: list[str]) -> str:
-        """Highest-ranked bump from ``bumps`` (``major > minor > patch``).
-
-        An empty list defaults to ``'patch'``.
-        """
-        if not bumps:
-            return "patch"
-        return max(bumps, key=lambda b: cls._BUMP_RANK.get(b, -1))
-
-    @staticmethod
-    def _merge_sections(fragments: list[dict[str, list[str]]]) -> dict[str, list[str]]:
-        """Merge multiple parsed fragments into a single section map.
-
-        Bullets from different fragments that share a section heading are
-        concatenated directly (no blank line between them) to match the
-        dominant style in IsaacLab's existing ``CHANGELOG.rst`` files.
-        """
-        merged: dict[str, list[str]] = {}
-        for frag in fragments:
-            for section, lines in frag.items():
-                if section not in merged:
-                    merged[section] = list(lines)
-                else:
-                    merged[section].extend(lines)
-        return merged
-
-    @classmethod
-    def _format_entry(cls, version: str, sections: dict[str, list[str]]) -> str:
-        """Return a complete RST version entry, ready to prepend to ``CHANGELOG.rst``.
-
-        Sections appear in :attr:`_SECTION_ORDER` (Added, Changed,
-        Deprecated, Removed, Fixed). Anything else keeps insertion order
-        *after* the canonical ones.
-        """
-        today = date.today().strftime("%Y-%m-%d")
-        heading = f"{version} ({today})"
-        out = [heading, "~" * len(heading), ""]
-
-        ordered = [s for s in cls._SECTION_ORDER if s in sections]
-        extras = [s for s in sections if s not in cls._SECTION_ORDER]
-
-        for section in ordered + extras:
-            out.append(section)
-            out.append("^" * len(section))
-            out.append("")
-            for line in sections[section]:
-                out.append(line.rstrip("\n"))
-            out.append("")
-
-        return "\n".join(out) + "\n"
-
-
-@dataclass(frozen=True)
-class Package:
-    """A source/<pkg>/ directory the changelog tool can manage.
-
-    A package is "managed" if it has both a version metadata file (see
-    :attr:`toml_path`, the file the compiler bumps) and a
-    ``docs/CHANGELOG.rst`` (the file the compiler updates).
-    :meth:`discover` returns only managed packages; instances created
-    directly may not be managed (use :attr:`is_managed`).
-    """
-
-    root: Path
-
-    @property
-    def name(self) -> str:
-        return self.root.name
-
-    @property
-    def changelog_path(self) -> Path:
-        return self.root / "docs" / "CHANGELOG.rst"
-
-    @property
-    def toml_path(self) -> Path:
-        return self.root / "pyproject.toml"
-
-    @property
-    def default_fragment_dir(self) -> Path:
-        return self.root / "changelog.d"
-
-    @property
-    def is_managed(self) -> bool:
-        return self.toml_path.is_file() and self.changelog_path.is_file()
-
-    def current_version(self) -> Version:
-        in_project = False
-        for line in self.toml_path.read_text(encoding="utf-8").splitlines():
-            if re.match(r"^\[project\]", line):
-                in_project = True
-            elif re.match(r"^\[", line):
-                in_project = False
-            if in_project:
-                m = re.match(r'^version\s*=\s*"([^"]+)"', line)
-                if m:
-                    return Version(m.group(1))
-        raise ValueError(f"{self.name}: no version field found under [project] in {self.toml_path}")
-
-    @classmethod
-    def declared_version(cls, root: Path) -> str | None:
-        """Return the version the package at ``root`` declares, or ``None`` if unreadable.
-
-        The tolerant counterpart to :meth:`current_version`, for callers
-        enumerating directories they do not control — :class:`LockFile`
-        walks every uv workspace member, including ones the changelog
-        compiler does not manage. A missing or malformed version metadata
-        file yields ``None`` instead of raising, so one unmanaged member
-        cannot fail an operation that spans the whole workspace.
-
-        Args:
-            root: The package directory (``source/<pkg>``).
-        """
-        pkg = cls(root)
-        if not pkg.toml_path.is_file():
-            return None
-        try:
-            return str(pkg.current_version())
-        except (OSError, ValueError):
-            return None
-
-    def write_changelog_entry(self, entry: str, *, dry_run: bool) -> list[Path]:
-        """Prepend ``entry`` to this package's CHANGELOG.rst. Returns the
-        list of paths written (empty in dry-run, so callers like
-        :class:`AutoBumpRun` get a single source of truth for "what just
-        changed on disk")."""
-        text = self.changelog_path.read_text(encoding="utf-8")
-        # Self-heal a header that lacks the trailing blank line. The compile
-        # regex needs ``Changelog\n---+\n\n`` as an anchor; a contributor who
-        # ships ``Changelog\n---+\n`` (the isaaclab_ppisp shape PR #5748
-        # introduced) would otherwise wedge the nightly. Insert the missing
-        # ``\n`` in-memory and write it back so the on-disk file ends up
-        # canonical on first compile. No-op when the blank line is already
-        # there (negative lookahead).
-        text = re.sub(r"^(Changelog\n-+)\n(?!\n)", r"\1\n\n", text, count=1, flags=re.MULTILINE)
-        m = CHANGELOG_HEADER_RE.search(text)
-        if not m:
-            raise ValueError(f"Could not locate changelog header in {self.changelog_path}")
-        updated = text[: m.end()] + entry + "\n" + text[m.end() :]
-        if dry_run:
-            print(f"\n{'=' * 60}")
-            print(f"DRY RUN — would write to {_display_path(self.changelog_path)}")
-            print(f"{'=' * 60}")
-            print(entry)
-            return []
-        self.changelog_path.write_text(updated, encoding="utf-8")
-        return [self.changelog_path]
-
-    def write_version(self, new_version: Version, *, dry_run: bool) -> list[Path]:
-        """Set ``version = "<new_version>"`` in this package's version metadata file.
-
-        Which file that is depends on the branch layout — :attr:`toml_path`
-        resolves it (``pyproject.toml`` here, ``config/extension.toml`` on
-        release branches that predate #6505). Returns the list of paths
-        written (empty in dry-run) so :class:`AutoBumpRun` has a single
-        source of truth for what changed on disk.
-        """
-        text = self.toml_path.read_text(encoding="utf-8")
-        in_project = False
-        new_lines = []
-        for line in text.splitlines(keepends=True):
-            if re.match(r"^\[project\]", line):
-                in_project = True
-            elif re.match(r"^\[", line):
-                in_project = False
-            if in_project and re.match(r'^version\s*=\s*"[^"]+"', line):
-                line = re.sub(r'^(version\s*=\s*)"[^"]+"', f'\\1"{new_version}"', line)
-            new_lines.append(line)
-        if dry_run:
-            print(f'DRY RUN — would set version = "{new_version}" in {_display_path(self.toml_path)}')
-            return []
-        self.toml_path.write_text("".join(new_lines), encoding="utf-8")
-        return [self.toml_path]
-
-    @classmethod
-    def from_name(cls, name: str, packages_root: Path = PACKAGES_ROOT) -> Package:
-        return cls(packages_root / name)
-
-    @classmethod
-    def discover(cls, packages_root: Path = PACKAGES_ROOT) -> list[Package]:
-        """Return all managed packages under ``packages_root``, sorted by name."""
-        if not packages_root.is_dir():
-            return []
-        return sorted(
-            (cls(child) for child in packages_root.iterdir() if child.is_dir() and cls(child).is_managed),
-            key=lambda p: p.name,
-        )
-
-    def compile(
-        self,
-        *,
-        fragments_dir: Path | None = None,
-        explicit_version: Version | None = None,
-        dry_run: bool = False,
-    ) -> tuple[bool, list[Path]]:
-        """Compile fragments for this package.
-
-        There are exactly two modes: ``dry_run=True`` previews and writes
-        nothing; ``dry_run=False`` writes the new entry, bumps the version,
-        **and** deletes the consumed fragments. There is deliberately no
-        third "write but keep fragments" mode — leaving fragments in place
-        after a real compile is a footgun (the next compile would re-emit
-        them as a duplicate version block).
-
-        Args:
-            fragments_dir: Read fragments from here instead of
-                :attr:`default_fragment_dir`. Useful for previewing against
-                example fixtures.
-            explicit_version: Pin the new version to this string (skips the
-                per-fragment bump inference).
-            dry_run: Preview only — no files are written or deleted.
-
-        Returns:
-            ``(compiled, touched)`` where ``compiled`` is ``True`` if at
-            least one fragment was found and processed, and ``touched`` is
-            the list of paths actually written to disk (empty in dry-run
-            mode). The ``touched`` list is the in-process manifest
-            :class:`AutoBumpRun` stages — no out-of-band file or glob
-            needed.
-        """
-        batch = FragmentBatch.from_dir(self._resolve_fragments_dir(fragments_dir))
-
-        for p in batch.invalid:
-            print(
-                f"  WARNING: {_display_path(p)} does not match any recognised fragment "
-                "pattern (<slug>.rst, <slug>.minor.rst, <slug>.major.rst, <slug>.skip) — skipping.",
-                file=sys.stderr,
-            )
-
-        if not batch.valid:
-            if batch.skip_paths:
-                n = len(batch.skip_paths)
-                if dry_run:
-                    print(f"  {self.name}: would clean {n} stale skip file(s).")
-                else:
-                    # No entry and no bump, but the skip files are still gone
-                    # from the working tree — report them so the deletion is
-                    # staged rather than silently reverted on next checkout.
-                    print(f"  {self.name}: cleaned {n} stale skip file(s).")
-                    return False, batch.delete_skips()
-            else:
-                print(f"  {self.name}: no fragments, skipping.")
-            return False, []
-
-        # Apply the same content-validation rules the PR gate uses, so a
-        # malformed fragment that somehow reached this package (e.g. a
-        # stale fragment that predates a content-rule tightening, or a
-        # locally-edited file) doesn't silently produce a half-empty
-        # version block. Runs every fragment that survived filename
-        # validation in ``from_dir``.
-        validation_errors = [(f, err) for f in batch.valid if (err := f.validate()) is not None]
-        if validation_errors:
-            for f, err in validation_errors:
-                print(f"  ERROR: {_display_path(f.path)}: {err}", file=sys.stderr)
-            raise ValueError(
-                f"{self.name}: {len(validation_errors)} fragment(s) failed content validation; "
-                "fix or remove them before compiling."
-            )
-
-        parsed_pairs = batch.parsed()
-        if not parsed_pairs:
-            print(f"  {self.name}: all fragments empty after parsing, skipping.")
-            return False, []
-
-        new_version, bump_label, entry = batch.compile_to_entry(
-            self.current_version(), explicit_version=explicit_version
-        )
-        print(f"  {self.name}: {len(parsed_pairs)} fragment(s) → version {new_version}{bump_label}")
-
-        if not self.changelog_path.exists():
-            # Should never happen with managed packages discovered via
-            # ``Package.discover()`` — defensive check for callers that
-            # construct a ``Package`` directly with an unmanaged root.
-            raise ValueError(
-                f"{_display_path(self.changelog_path)} does not exist; "
-                f"package {self.name!r} is not managed (missing CHANGELOG.rst)."
-            )
-        touched: list[Path] = []
-        touched.extend(self.write_changelog_entry(entry, dry_run=dry_run))
-        touched.extend(self.write_version(new_version, dry_run=dry_run))
-
-        if not dry_run:
-            deleted_frags, deleted_skips = batch.delete_all()
-            # Deletions are part of the change set: they must be staged with
-            # the entry that consumed them, or the fragments come back on the
-            # next checkout and recompile into a duplicate version block.
-            touched.extend(deleted_frags)
-            touched.extend(deleted_skips)
-            msg = f"  {self.name}: deleted {len(deleted_frags)} fragment(s)"
-            if deleted_skips:
-                msg += f" and {len(deleted_skips)} skip file(s)"
-            print(msg + ".")
-
-        return True, touched
-
-    def _resolve_fragments_dir(self, override: Path | None) -> Path:
-        """Pick the directory ``compile`` should read fragments from.
-
-        ``None`` means "use this package's own ``changelog.d/``"; an
-        absolute path is used as-is; a relative path is resolved against
-        ``REPO_ROOT`` so callers can pass things like
-        ``tools/changelog/test/integration/01_patch_bump/fragments`` without
-        worrying about the cwd.
-        """
-        if override is None:
-            return self.default_fragment_dir
-        return override if override.is_absolute() else (REPO_ROOT / override).resolve()
-
-
-@dataclass(frozen=True)
-class PRDiff:
-    """A snapshot of "what this PR changed against its base branch."
-
-    Wraps two views from the same git diff: ``changed`` (any file modified
-    or added) and ``added`` (the strict subset that's new on this branch).
-    Tests construct ``PRDiff`` directly with synthetic sets;
-    :meth:`from_git` runs the real ``git diff`` for production use.
-    """
-
-    changed: set[str]
-    added: set[str]
-
-    @classmethod
-    def from_git(cls, base_ref: str) -> PRDiff:
-        """Run ``git diff`` against ``origin/<base_ref>...HEAD`` to populate the diff."""
-
-        def _diff(extra_args: list[str]) -> set[str]:
-            result = subprocess.run(
-                ["git", "diff", "--name-only", *extra_args, f"origin/{base_ref}...HEAD"],
-                capture_output=True,
-                text=True,
-                check=True,
-                cwd=REPO_ROOT,
-            )
-            return {f for f in result.stdout.splitlines() if f}
-
-        return cls(changed=_diff([]), added=_diff(["--diff-filter=A"]))
-
-    def evaluate(
-        self,
-        packages: list[Package],
-    ) -> tuple[list[str], list[tuple[str, str]]]:
-        """Apply the PR-gate rules and return ``(missing_packages, invalid_fragments)``.
-
-        Rules:
-
-        1. **Immutability** — every fragment file in the diff must be in
-           ``added`` (added on this branch). Modifying or renaming an existing
-           fragment is rejected with a hint to add a new one instead.
-
-        2. **Content validity** — every added ``*.rst`` fragment must parse
-           (recognised section headings + at least one bullet). ``.skip`` and
-           ``.gitkeep`` are exempt.
-
-        3. **Slug uniqueness** — within a package's ``changelog.d/``, no two
-           fragments may share the same slug. If an added fragment's slug
-           collides with an existing or co-added fragment, fail with a hint
-           to rename (e.g. append ``-2``).
-
-        4. **Required fragment per touched package** — for each managed
-           package the PR touches in ``source/`` (outside ``changelog.d/``),
-           the PR must *add* at least one valid fragment to that package's
-           ``changelog.d/``. Chained PRs (parent PR's fragment shows up in
-           the child's diff) naturally satisfy this — slug uniqueness is
-           the only constraint that matters.
-        """
-        missing: list[str] = []
-        invalid_fragments: list[tuple[str, str]] = []
-
-        for pkg in packages:
-            pkg_prefix = f"source/{pkg.name}/"
-            changelog_dir = f"source/{pkg.name}/changelog.d/"
-
-            source_changed = [f for f in self.changed if f.startswith(pkg_prefix) and not f.startswith(changelog_dir)]
-            fragment_changes = [f for f in self.changed if f.startswith(changelog_dir)]
-
-            # Map *pre-existing* fragments in the package's changelog.d/ by slug,
-            # for the uniqueness check below. The CI checkout contains both
-            # base-branch fragments and the PR's additions side by side, so we
-            # must explicitly exclude added files — otherwise an added file can
-            # overwrite the entry for a colliding pre-existing fragment with
-            # the same slug, hiding the very collision we're trying to detect.
-            # Skip ``.gitkeep`` and unrecognised filenames — they can't collide.
-            added_basenames = {Path(f).name for f in self.added if f.startswith(changelog_dir)}
-            existing_slugs: dict[str, str] = {}
-            existing_dir = pkg.default_fragment_dir
-            if existing_dir.is_dir():
-                for p in existing_dir.iterdir():
-                    if p.is_dir() or p.name == ".gitkeep" or p.name in added_basenames:
-                        continue
-                    slug = Fragment.parse_slug(p.name)
-                    if slug is not None:
-                        existing_slugs[slug] = p.name
-
-            added_slugs: dict[str, str] = {}
-            for f in fragment_changes:
-                path = Path(f)
-                if path.name == ".gitkeep":
-                    continue
-
-                # Rule 1: immutability — modifying an existing fragment is forbidden.
-                if f not in self.added:
-                    invalid_fragments.append(
-                        (
-                            f,
-                            "fragments are immutable — add a new fragment with a different slug "
-                            "instead of editing an existing one",
-                        )
-                    )
-                    continue
-
-                # Rule 2: content validity (only for *.rst, not *.skip).
-                if not SKIP_RE.match(path.name):
-                    err = Fragment(REPO_ROOT / f).validate()
-                    if err:
-                        invalid_fragments.append((f, err))
-                        continue
-
-                # Rule 3: slug uniqueness within the package's changelog.d/.
-                slug = Fragment.parse_slug(path.name)
-                if slug is None:
-                    # Filename validation already flagged this above for *.rst,
-                    # but a malformed *.skip would slip through. Surface it.
-                    invalid_fragments.append(
-                        (f, "invalid filename — must be <slug>.rst, <slug>.minor.rst, <slug>.major.rst, or <slug>.skip")
-                    )
-                    continue
-                if slug in existing_slugs and existing_slugs[slug] != path.name:
-                    invalid_fragments.append(
-                        (
-                            f,
-                            f"slug {slug!r} collides with existing fragment "
-                            f"{existing_slugs[slug]!r} — rename to {slug}-2 (or any unused slug)",
-                        )
-                    )
-                    continue
-                if slug in added_slugs and added_slugs[slug] != path.name:
-                    invalid_fragments.append(
-                        (
-                            f,
-                            f"slug {slug!r} collides with another added fragment "
-                            f"{added_slugs[slug]!r} — rename one to {slug}-2 (or any unused slug)",
-                        )
-                    )
-                    continue
-                added_slugs[slug] = path.name
-
-            if not source_changed:
-                continue
-
-            # Rule 4: this PR must add at least one valid fragment for the package.
-            owned = [
-                f
-                for f in fragment_changes
-                if f in self.added and (FRAGMENT_RE.match(Path(f).name) or SKIP_RE.match(Path(f).name))
-            ]
-            if not owned:
-                missing.append(pkg.name)
-
-        return missing, invalid_fragments
-
+# Re-exported so ``import cli`` stays the one entry point for the tool and
+# its tests: the domain model moved to its own module for readability, not
+# to change who may reach for it.
+from packages import (  # noqa: F401
+    CHANGELOG_HEADER_RE,
+    FRAGMENT_RE,
+    PACKAGES_ROOT,
+    REPO_ROOT,
+    SKIP_RE,
+    Fragment,
+    FragmentBatch,
+    Package,
+    PRDiff,
+    Version,
+    _display_path,
+)
 
 # ---------------------------------------------------------------------------
 # Subcommand handlers
@@ -1104,294 +259,6 @@ def cmd_check(args: argparse.Namespace, _parser: argparse.ArgumentParser) -> int
 
     print("✓ All modified packages have valid changelog fragments.")
     return 0
-
-
-# ---------------------------------------------------------------------------
-# Nightly auto-bump: compile + stage + commit + push, all in-process
-# ---------------------------------------------------------------------------
-
-
-class GitError(Exception):
-    """Base class for failures raised by :class:`GitRepo`."""
-
-
-class GitRepo:
-    """Thin subprocess wrapper around the ``git`` CLI scoped to one working tree.
-
-    Owns only the working directory and the policy decisions that need
-    typed errors. All other behavior delegates straight to ``git``, so
-    the unit tests can run against a real tempdir repo + bare-repo remote
-    instead of mocking subprocess.
-    """
-
-    class NonFastForward(GitError):
-        """Raised when ``git push`` is rejected because the remote advanced."""
-
-    def __init__(self, cwd: Path):
-        self.cwd = cwd
-
-    def _run(self, *args: str, check: bool = True) -> subprocess.CompletedProcess[str]:
-        # ``-c color.ui=false`` because this output is parsed, not displayed.
-        # git suppresses colour for a pipe by default, but a developer or
-        # runner with ``color.ui = always`` configured overrides that, and the
-        # resulting ANSI escapes silently break line-prefix matching: a diff
-        # line arrives as ``\x1b[31m-version = "1.2.3"``, so the version
-        # extraction reports "?" and the non-fast-forward detection stops
-        # recognising a rejected push. Neither failure is loud.
-        return subprocess.run(
-            ["git", "-c", "color.ui=false", *args],
-            cwd=self.cwd,
-            text=True,
-            capture_output=True,
-            check=check,
-        )
-
-    def config(self, key: str, value: str) -> None:
-        self._run("config", key, value)
-
-    def add(self, paths: Iterable[Path | str]) -> None:
-        """Stage ``paths``, including deletions.
-
-        ``git add`` on a path that is gone *and* was never tracked aborts the
-        whole invocation with ``pathspec ... did not match any files``, which
-        would take the entire staging step down over one stray file. Deleted
-        paths are therefore filtered to the ones git actually knows about;
-        a deleted-but-tracked path stages its removal, as intended.
-        """
-        path_strs = [str(p) for p in paths]
-        vanished = [p for p in path_strs if not Path(p).exists()]
-        if vanished:
-            known = set(self._tracked(vanished))
-            path_strs = [p for p in path_strs if Path(p).exists() or p in known]
-        if not path_strs:
-            return
-        self._run("add", "--", *path_strs)
-
-    def _tracked(self, paths: Iterable[Path | str]) -> list[str]:
-        """Return the subset of ``paths`` git has under version control."""
-        listed = self._run("ls-files", "--", *[str(p) for p in paths]).stdout.split("\n")
-        # ``ls-files`` reports repo-relative paths; callers hold absolute ones.
-        return [str(self.cwd / line) for line in listed if line]
-
-    def has_staged_changes(self) -> bool:
-        return self._run("diff", "--staged", "--quiet", check=False).returncode != 0
-
-    def staged_diff(self, path: Path | str) -> str:
-        return self._run("diff", "--staged", "--", str(path)).stdout
-
-    def commit(self, message: str) -> None:
-        self._run("commit", "-m", message)
-
-    def fetch(self, remote: str, ref: str) -> None:
-        self._run("fetch", remote, ref)
-
-    def rebase(self, onto: str) -> None:
-        self._run("rebase", onto)
-
-    def push(self, remote: str, refspec: str) -> None:
-        result = self._run("push", remote, refspec, check=False)
-        if result.returncode == 0:
-            return
-        stderr = (result.stderr or "") + (result.stdout or "")
-        # ``non-fast-forward`` is the git-side message; ``[rejected]`` covers
-        # the GitHub ruleset path. Either one means the remote tip moved or
-        # rejected the push for a recoverable reason; let the orchestrator
-        # decide whether to retry.
-        if "non-fast-forward" in stderr or "[rejected]" in stderr:
-            raise self.NonFastForward(stderr.strip())
-        raise GitError(f"git push failed: {stderr.strip()}")
-
-
-class AutoBumpRun:
-    """One-shot orchestrator for the nightly auto-commit lifecycle.
-
-    Compiles every managed package, stages whatever the compile actually
-    wrote (no external glob — the touched-paths list comes from each
-    :meth:`Package.compile` return value), builds a bot-attributed commit,
-    and pushes it to the target branch. Retries the push on
-    non-fast-forward by fetching and rebasing the auto-commit onto the new
-    tip, so a human commit landing mid-run doesn't waste the batch.
-
-    Identity is hardcoded — the bot's user name and email are deterministic
-    public values derived from the GitHub App registration (anyone reading
-    a nightly commit can see them). The App credential itself stays in repo
-    secrets and is consumed by the workflow, not this class.
-    """
-
-    AUTHOR_NAME = "isaaclab-bot[bot]"
-    AUTHOR_EMAIL = "282401363+isaaclab-bot[bot]@users.noreply.github.com"
-    PUSH_RETRIES = 3
-    COMMIT_PREFIX = "[CI][Auto Version Bump]"
-
-    def __init__(
-        self,
-        *,
-        branch: str,
-        remote: str,
-        event_name: str = "manual",
-        dry_run: bool = False,
-        repo_root: Path = REPO_ROOT,
-    ):
-        self.branch = branch
-        self.remote = remote
-        self.event_name = event_name
-        self.dry_run = dry_run
-        self.repo_root = repo_root
-        self.repo = GitRepo(repo_root)
-        self.packages = Package.discover(packages_root=repo_root / "source")
-        self.lock = LockFile(repo_root, Package.declared_version)
-        self.touched: list[Path] = []
-        self.failures: list[tuple[str, str]] = []
-        self.any_compiled = False
-
-    def run(self) -> int:
-        self._compile_all()
-        if self.dry_run:
-            self._sync_lock()
-            self._report_dry_run()
-            return self._exit_code()
-        if not self.touched:
-            if not self.any_compiled:
-                print("No fragments found in any package.")
-            else:
-                print("All compiles ran but produced no on-disk writes (already up to date).")
-            return self._exit_code()
-        self._sync_lock()
-        self._stage_and_commit()
-        self._push_with_retry()
-        return self._exit_code()
-
-    def _compile_all(self) -> None:
-        for pkg in self.packages:
-            try:
-                compiled, touched = pkg.compile(dry_run=self.dry_run)
-            except (FileNotFoundError, ValueError) as e:
-                print(f"  ERROR ({pkg.name}): {e}", file=sys.stderr)
-                self.failures.append((pkg.name, str(e)))
-                continue
-            self.any_compiled = self.any_compiled or compiled
-            self.touched.extend(touched)
-
-    def _sync_lock(self) -> None:
-        """Re-point ``uv.lock`` at the versions the compile just wrote.
-
-        Runs once, after every package has compiled — ``uv.lock`` is a
-        single repo-level artifact pinning all members at once, so there is
-        nothing per-package about it. The written path joins
-        :attr:`touched`, which is what gets the lock staged through the same
-        manifest as every other compiler output; no separate ``git add``
-        exists to drift.
-
-        A failure is recorded like a failed package compile — non-zero exit,
-        red tile — but deliberately does not block the commit. A lock that
-        needs a full ``uv lock`` was already stale before this run started,
-        and wedging the nightly over it would strand every package's
-        changelog for a problem the auto-commit did not cause.
-        """
-        try:
-            self.touched.extend(self.lock.sync(dry_run=self.dry_run))
-        except LockFile.Error as e:
-            print(f"  ERROR ({LockFile.LOCK_NAME}): {e}", file=sys.stderr)
-            self.failures.append((LockFile.LOCK_NAME, str(e)))
-
-    def _report_dry_run(self) -> None:
-        if self.any_compiled:
-            print(f"DRY RUN — compile complete; would commit/push to {self.branch}.")
-        else:
-            print("DRY RUN — no fragments found in any package.")
-
-    def _stage_and_commit(self) -> None:
-        # Author identity belongs in-process: the workflow YAML stops
-        # carrying changelog-tool knowledge so cli.py-only PRs don't need
-        # paired YAML edits to stay consistent.
-        self.repo.config("user.name", self.AUTHOR_NAME)
-        self.repo.config("user.email", self.AUTHOR_EMAIL)
-        self.repo.add(self.touched)
-        if not self.repo.has_staged_changes():
-            print("Nothing actually staged after compile — skipping commit.")
-            return
-        self.repo.commit(self._build_commit_message())
-        print(f"Committed bump for {len(self.touched)} file(s).")
-
-    def _relative(self, path: Path) -> Path:
-        """``path`` relative to the repo root, or unchanged if it lies outside."""
-        try:
-            return path.relative_to(self.repo_root)
-        except ValueError:
-            return path
-
-    def _build_commit_message(self) -> str:
-        # Derive the per-package "old → new" lines from the staged version
-        # metadata diffs. The set of version files is taken from the packages
-        # themselves (:attr:`Package.toml_path`) rather than a hardcoded
-        # filename, so this produces a correct message both here — where
-        # versions live in ``pyproject.toml`` — and on release branches that
-        # still keep them in ``config/extension.toml``. Files touched by a
-        # future write site that carries no ``version`` line (``uv.lock``,
-        # ``CHANGELOG.rst``) are staged but not enumerated here.
-        lines = [
-            f"{self.COMMIT_PREFIX} Compile changelog fragments ({self.event_name})",
-            "",
-            "Bumped packages:",
-        ]
-        version_files = {pkg.toml_path: pkg.name for pkg in self.packages}
-        for path in sorted(p for p in self.touched if p in version_files):
-            diff = self.repo.staged_diff(self._relative(path))
-            old = _extract_version_from_diff(diff, "-")
-            new = _extract_version_from_diff(diff, "+")
-            lines.append(f"- {version_files[path]}: {old} → {new}")
-        return "\n".join(lines) + "\n"
-
-    def _push_with_retry(self) -> None:
-        refspec = f"HEAD:refs/heads/{self.branch}"
-        last_err: GitRepo.NonFastForward | None = None
-        for attempt in range(self.PUSH_RETRIES):
-            try:
-                self.repo.push(self.remote, refspec)
-                if attempt > 0:
-                    print(f"Push succeeded on attempt {attempt + 1}.")
-                else:
-                    print(f"Push succeeded → {self.remote} {refspec}.")
-                return
-            except GitRepo.NonFastForward as e:
-                last_err = e
-                if attempt + 1 == self.PUSH_RETRIES:
-                    break
-                print(
-                    f"  push rejected (attempt {attempt + 1}/{self.PUSH_RETRIES}); "
-                    f"fetching and rebasing onto {self.branch}: {e}",
-                    file=sys.stderr,
-                )
-                self.repo.fetch(self.remote, self.branch)
-                self.repo.rebase("FETCH_HEAD")
-        # All retries exhausted.
-        assert last_err is not None  # the loop only exits here when it raised at least once
-        raise GitError(f"push failed after {self.PUSH_RETRIES} attempts; last error: {last_err}")
-
-    def _exit_code(self) -> int:
-        if self.failures:
-            print(file=sys.stderr)
-            print(f"::error::{len(self.failures)} package(s) failed to compile:", file=sys.stderr)
-            for name, reason in self.failures:
-                print(f"  • {name}: {reason}", file=sys.stderr)
-            return 1
-        return 0
-
-
-def _extract_version_from_diff(diff_text: str, prefix: str) -> str:
-    """Pull the version string out of a staged version-metadata diff line.
-
-    The diff contains lines like ``-version = "1.2.3"`` (old) and
-    ``+version = "1.3.0"`` (new). ``prefix`` selects which side to read.
-    Returns ``"?"`` if no match — the commit message is informational, not
-    machine-parsed, so a missing value shouldn't fail the run.
-    """
-    for line in diff_text.splitlines():
-        if line.startswith(f"{prefix}version"):
-            m = re.search(r'"([^"]+)"', line)
-            if m:
-                return m.group(1)
-    return "?"
 
 
 def cmd_auto_bump(args: argparse.Namespace, _parser: argparse.ArgumentParser) -> int:
